requirements.txt
python-telegram-bot[all]>=20.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
openai>=1.0.0
